requirements.txt
tensorflow~=2.19.0
opencv-python~=4.11.0.86
numpy~=2.1.3
matplotlib
scikit-learn

pillow~=11.2.1
pyttsx3~=2.98